+++ CotEditor/Sources/CEPrefEncodingDataSource.h
@@ -5,12 +5,12 @@
  CotEditor
  http://coteditor.com
  
- Created on 2004-12-16 by 1024jp
+ Created on 2014-12-16 by 1024jp
  encoding="UTF-8"
  ------------------------------------------------------------------------------
  
  © 2004-2007 nakamuxu
- © 2014 1024jp
+ © 2014-2015 1024jp
  
  This program is free software; you can redistribute it and/or modify it under
  the terms of the GNU General Public License as published by the Free Software
@@ -34,7 +34,6 @@
 @interface CEPrefEncodingDataSource : NSObject <NSTableViewDataSource, NSTabViewDelegate>
 
 // Public method
-- (void)setupEncodingsToEdit;
-- (void)writeEncodingsToUserDefaults;
+- (void)writeToUserDefaults;
 
 @end