+++ CotEditor/Sources/CETheme.h
@@ -70,4 +70,7 @@
 /// designated initializer
 - (nullable instancetype)initWithName:(nonnull NSString *)themeName NS_DESIGNATED_INITIALIZER;
 
+// unavailable initializer
+- (nullable instancetype)init __attribute__((unavailable("use -initWithName: instead")));
+
 @end